Update patch set 1
Patch Set 1:

Hmm I dunno, I'm not across the structure of the layouts, but if you want something on every page it seems like it should be in a template header or footer? I presume this default layout may not necessarily be used everywhere...

PS: I don't think it makes any difference whether the script itself is included at the top or bottom of the page (it was already in the head.html), does it?

Patch-set: 1
Reviewer: Gerrit User 1000082 <1000082@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Verified=0